Voice-transcript/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
attrs==26.1.0
boto3==1.42.88
botocore==1.42.88
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.135.3
frozenlist==1.8.0
greenlet==3.4.0
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.1.0
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
numpy==2.4.4
pandas==3.0.2
passlib==1.7.4
propcache==0.4.1
pyasn1==0.6.3
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
requests==2.33.1
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlmodel==0.0.38
starlette==1.0.0
twilio==9.10.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
yarl==1.23.0
